pyproject.toml
[project]
name = "docuview"
version = "0.1.0"
description = "Document parser service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "alembic>=1.13",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "python-dotenv>=1.0",
    "honcho>=1.1",
    "ruff>=0.4",
    "black>=24.4",
    "pytest-cov>=5.0",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pre-commit>=3.7",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docuview"]

[tool.pytest.ini_options]
pythonpath = ["src"]


testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
]
ignore = [
    "E501",  # line too long — black/formatter следит
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = true
warn_unused_ignores = true
# НЕ используем глобальный ignore_missing_imports — он конфликтует с
# warn_unused_ignores: mypy подавляет импорт-ошибку, потом видит
# оставшийся # type: ignore как лишний и снова падает.
# types-PyYAML установлен в dependency-groups.dev, поэтому yaml
# покрыт стабами. Отключаем import-untyped глобально для остальных
# пакетов без py.typed (magic, и др.).
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = ["magic"]
ignore_missing_imports = true

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20260408",
]
